requirements.txt
Django==1.11.3
django-allauth==0.32.0
django-comments-xtd==2.0.5
django-crispy-forms==1.6.1
django-jinja==2.3.1
dj-database-url==0.4.2
django-precise-bbcode==1.2.8
django-taggit==0.22.1
psycopg2==2.7.1
